manage-sensors-app/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
Django==5.2.1
django-redis==6.0.0
django-tables2==2.7.5
idna==3.10
narwhals==1.41.0
packaging==25.0
paho-mqtt==2.1.0
plotly==6.1.2
psycopg2-binary==2.9.10
pycparser==2.22
python-dotenv==1.1.1
redis==6.2.0
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
